Chat notif fix (#66)
* fixed notification not updating

* fixed get error being logged

---------

+++ src/app/Components/Chat/chat-area/chat-area.component.ts
@@ -10,17 +10,18 @@ import {
   Output,
   ViewChild,
 } from '@angular/core';
-import { MessageDto } from '../../../Dtos/Message/MessageDto';
-import { DateHelper } from '../../../Helpers/DateHelper';
-import { MessageStatus } from '../../../Dtos/Message/MessageStatus';
-import { SignalRService } from '../../../Services/SignalR/signal-r.service';
-import { ActivatedRoute } from '@angular/router';
-import { Modal } from 'bootstrap';
-import { DeleteMessageRequestDto } from '../../../Dtos/Message/DeleteMessageRequestDto';
-import { InsertReportDto } from '../../../Dtos/Reports/InsertReportDto';
-import { ReportsService } from '../../../Services/Reports/reports.service';
-import { ChatService } from '../../../Services/Chat/chat.service';
-import { NotificationService } from '../../../Services/Chat/notification.service';
+import {MessageDto} from '../../../Dtos/Message/MessageDto';
+import {DateHelper} from '../../../Helpers/DateHelper';
+import {MessageStatus} from '../../../Dtos/Message/MessageStatus';
+import {SignalRService} from '../../../Services/SignalR/signal-r.service';
+import {ActivatedRoute} from '@angular/router';
+import {Modal} from 'bootstrap';
+import {DeleteMessageRequestDto} from '../../../Dtos/Message/DeleteMessageRequestDto';
+import {InsertReportDto} from '../../../Dtos/Reports/InsertReportDto';
+import {ReportsService} from '../../../Services/Reports/reports.service';
+import {ChatService} from '../../../Services/Chat/chat.service';
+import {NotificationService} from '../../../Services/Chat/notification.service';
+import {ChatUsersService} from "../../../Services/chat-users.service";
 
 @Component({
   selector: 'app-chat-area',
@@ -28,7 +29,7 @@ import { NotificationService } from '../../../Services/Chat/notification.service
   styleUrls: ['./chat-area.component.css'],
 })
 export class ChatAreaComponent
-  implements AfterViewChecked, OnInit, AfterViewInit, OnDestroy
+  implements OnInit, AfterViewInit, OnDestroy
 {
   public messages: MessageDto[] = [];
   public currentUser = '';
@@ -50,14 +51,19 @@ export class ChatAreaComponent
     private activeRoute: ActivatedRoute,
     private reportsService: ReportsService,
     private chatService: ChatService,
-    private notificationService: NotificationService
-  ) {}
-
-  ngOnDestroy(): void {
-    this.currentUser = '';
+    private chatUsersService: ChatUsersService
+  )
+  {
   }
 
-  ngAfterViewInit(): void {
+  ngOnDestroy(): void
+  {
+    this.currentUser = '';
+    this.chatUsersService.setData(null)
+  }
+
+  ngAfterViewInit(): void
+  {
     this.deleteModal = new Modal(
       document.getElementById('deleteStaticBackdrop')!
     );
@@ -66,36 +72,54 @@ export class ChatAreaComponent
     );
   }
 
-  ngOnInit(): void {
+  ngOnInit(): void
+  {
     this.checkConnection();
 
-    this.activeRoute.params.subscribe((params) => {
+    this.activeRoute.params.subscribe((params) =>
+    {
       this.currentUser = params['id'];
-      this.chatService.chatOpened.emit(this.currentUser);
-      this.getChat(20);
+
+
+        this.chatUsersService.getData().subscribe(data =>
+        {
+          if (data != null && this.currentUser != '')
+          {
+            console.log('user', this.currentUser.charCodeAt(0))
+            this.chatService.chatOpened.emit(this.currentUser);
+            console.log("hhhhhhhhhhhhhhhh")
+            this.getChat(20);
+          }
+        })
     });
 
     this.signalRService.newMessageEvent.subscribe({
-      next: (message: MessageDto) => {
-        if (message.senderId == this.currentUser) {
+      next: (message: MessageDto) =>
+      {
+        if (message.senderId == this.currentUser)
+        {
           message.messageStatus = MessageStatus.Seen;
           this.messages.push(message);
           this.signalRService.setMessageSeen(message.id, message.senderId);
         }
 
-        this.chatService.newMessage.emit({ message, user: message.senderId });
+        this.chatService.newMessage.emit({message, user: message.senderId});
       },
       error: (err: any) => console.error(err),
     });
 
     this.signalRService.messageSeenEvent.subscribe({
-      next: (messageId: string) => {
+      next: (messageId: string) =>
+      {
         debugger;
         let msg = this.messages.find((m) => m.id == messageId);
-        if (msg) {
+        if (msg)
+        {
           msg.messageStatus = MessageStatus.Seen;
           this.messages = [...this.messages];
-        } else {
+        }
+        else
+        {
           console.error('No message with the received id');
         }
       },
@@ -103,9 +127,12 @@ export class ChatAreaComponent
     });
 
     this.signalRService.allMessagesSeenEvent.subscribe({
-      next: (userId: string) => {
-        if (userId == this.currentUser) {
-          for (let message of this.messages) {
+      next: (userId: string) =>
+      {
+        if (userId == this.currentUser)
+        {
+          for (let message of this.messages)
+          {
             message.messageStatus = MessageStatus.Seen;
           }
           this.messages = [...this.messages];
@@ -114,16 +141,19 @@ export class ChatAreaComponent
     });
   }
 
-  ngAfterViewChecked(): void {}
 
-  checkConnection() {
-    if (!this.signalRService.isConnected) {
+  checkConnection()
+  {
+    if (!this.signalRService.isConnected)
+    {
       this.signalRService.startConnection();
     }
   }
 
-  checkNewDate(i: number): boolean {
-    if (i == 0) {
+  checkNewDate(i: number): boolean
+  {
+    if (i == 0)
+    {
       return true;
     }
 
@@ -137,14 +167,18 @@ export class ChatAreaComponent
     );
   }
 
-  sendMessage(message: HTMLInputElement) {
-    if (message.value != '' && this.currentUser != '') {
+  sendMessage(message: HTMLInputElement)
+  {
+    if (message.value != '' && this.currentUser != '')
+    {
       let date = new Date();
 
       this.signalRService
         .sendMessage(message.value, this.currentUser, date)
-        .then((messageId: string) => {
-          if (messageId != '') {
+        .then((messageId: string) =>
+        {
+          if (messageId != '')
+          {
             let msg = Object.assign(
               {},
               new MessageDto(
@@ -170,9 +204,12 @@ export class ChatAreaComponent
     }
   }
 
-  scrollToTheEnd() {
-    setTimeout(() => {
-      if (this.messagesDiv) {
+  scrollToTheEnd()
+  {
+    setTimeout(() =>
+    {
+      if (this.messagesDiv)
+      {
         this.messagesDiv.nativeElement.scrollTo({
           top: Number(this.messagesDiv.nativeElement.scrollHeight),
           left: 0,
@@ -182,11 +219,13 @@ export class ChatAreaComponent
     }, 200);
   }
 
-  getChat(numOfDays: number = 20) {
+  getChat(numOfDays: number = 20)
+  {
     this.chatService
       .GetChat(this.currentUser, numOfDays, new Date())
       .subscribe({
-        next: (data) => {
+        next: (data) =>
+        {
           this.messages = data;
           this.scrollToTheEnd();
         },
@@ -194,14 +233,16 @@ export class ChatAreaComponent
       });
   }
 
-  deleteMessage(messageId: string) {
+  deleteMessage(messageId: string)
+  {
     this.deletedMessageId = messageId;
     this.deleteContent =
       this.messages.find((m) => m.id == messageId)?.message ?? '';
     this.deleteModal.show();
   }
 
-  reportMessage(messageId: string) {
+  reportMessage(messageId: string)
+  {
     this.reportedMessageId = messageId;
     this.reportContentA = `You are reporting the following message:`;
     this.reportContentB =
@@ -209,11 +250,13 @@ export class ChatAreaComponent
     this.reportModal.show();
   }
 
-  confirmedDeleteMessage() {
+  confirmedDeleteMessage()
+  {
     this.chatService
       .DeleteMessage(new DeleteMessageRequestDto(this.deletedMessageId, null))
       .subscribe({
-        next: () => {
+        next: () =>
+        {
           this.submitted = true;
           this.success = true;
 
@@ -223,7 +266,8 @@ export class ChatAreaComponent
           this.messages.splice(index, 1);
           this.cancelDelete();
         },
-        error: (err) => {
+        error: (err) =>
+        {
           this.submitted = true;
           this.success = false;
 
@@ -232,7 +276,8 @@ export class ChatAreaComponent
       });
   }
 
-  confirmedReportMessage(obj: any) {
+  confirmedReportMessage(obj: any)
+  {
     let report = new InsertReportDto(
       obj.title,
       obj.statement,
@@ -242,12 +287,14 @@ export class ChatAreaComponent
       null
     );
     this.reportsService.insertReport(report).subscribe({
-      next: (): void => {
+      next: (): void =>
+      {
         this.submitted = true;
         this.success = true;
         this.cancelReport();
       },
-      error: (err) => {
+      error: (err) =>
+      {
         this.submitted = true;
         this.success = false;
         console.error(err);
@@ -255,13 +302,15 @@ export class ChatAreaComponent
     });
   }
 
-  cancelReport() {
+  cancelReport()
+  {
     this.reportContentA = '';
     this.reportContentB = '';
     this.reportedMessageId = '';
   }
 
-  cancelDelete() {
+  cancelDelete()
+  {
     this.deletedMessageId = '';
     this.deleteContent = '';
   }
